requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
elastic-transport==8.17.1
elasticsearch==8.13.0
Faker==37.4.2
idna==3.10
mysql-connector-python==9.4.0
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
